@@ -1,11 +1,11 @@
 name: cuberto_bottom_bar
-description: An animated bottom navigation bar supporting drawer icon at start or end or can be used with the drawer icon and is provided with 2 different style.
-version: 1.0.0
+description: An animated bottom navigation bar supporting drawer icon at start or end or can be used with the drawer icon and is provided with 2 different styles.
+version: 1.1.0
 author: Kushal Mahapatro  <kushalmahapatro@gmail.com>
 homepage: https://github.com/kushalmahapatro/cuberto_bottom_bar
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.5.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -15,10 +15,6 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-# For information on the generic Dart part of this file, see the
-# following page: https://www.dartlang.org/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
 
   # To add assets to your package, add an assets section, like this:
