@@ -15,25 +15,37 @@
 #include "FlutterPlugin.h"
 #include "FlutterTexture.h"
 
+NS_ASSUME_NONNULL_BEGIN
+
 @class FlutterEngine;
+
+/**
+ * The name used for semantic update notifications via `NSNotificationCenter`.
+ *
+ * The object passed as the sender is the `FlutterViewController` associated
+ * with the update.
+ */
+FLUTTER_EXPORT
+extern NSNotificationName const FlutterSemanticsUpdateNotification;
 
 /**
  * A `UIViewController` implementation for Flutter views.
  *
- * Dart execution, channel communication, texture registration, and plugin registration
- * are all handled by `FlutterEngine`.  Calls on this class to those members all proxy
- * through to the `FlutterEngine` attached FlutterViewController.
+ * Dart execution, channel communication, texture registration, and plugin registration are all
+ * handled by `FlutterEngine`. Calls on this class to those members all proxy through to the
+ * `FlutterEngine` attached FlutterViewController.
  *
- * A FlutterViewController can be initialized either with an already-running `FlutterEngine`,
- * or it can be initialized with a `FlutterDartProject` that will be used to spin up
- * a new `FlutterEngine`.  Developers looking to present and hide FlutterViewControllers
- * in native iOS applications will usually want to maintain the `FlutterEngine` instance
- * so as not to lose Dart-related state and asynchronous tasks when navigating back and
- * forth between a FlutterViewController and other `UIViewController`s.
+ * A FlutterViewController can be initialized either with an already-running `FlutterEngine` via
+ * the `initWithEngine:` initializer, or it can be initialized with a `FlutterDartProject` that
+ * will be used to implicitly spin up a new `FlutterEngine`. Creating a `FlutterEngine before
+ * showing a `FlutterViewController` can be used to pre-initialize the Dart VM and to prepare the
+ * isolate in order to reduce the latency to the first rendered frame. Holding a `FlutterEngine`
+ * independently of FlutterViewControllers can also be used to not to lose Dart-related state and
+ * asynchronous tasks when navigating back and forth between a FlutterViewController and other
+ * `UIViewController`s.
  */
 FLUTTER_EXPORT
-@interface FlutterViewController
-    : UIViewController <FlutterBinaryMessenger, FlutterTextureRegistry, FlutterPluginRegistry>
+@interface FlutterViewController : UIViewController <FlutterTextureRegistry, FlutterPluginRegistry>
 
 /**
  * Initializes this FlutterViewController with the specified `FlutterEngine`.
@@ -41,24 +53,24 @@ FLUTTER_EXPORT
  * The initialized viewcontroller will attach itself to the engine as part of this process.
  *
  * @param engine The `FlutterEngine` instance to attach to.
- * @param nibNameOrNil The NIB name to initialize this UIViewController with.
- * @param nibBundleOrNil The NIB bundle.
+ * @param nibName The NIB name to initialize this UIViewController with.
+ * @param nibBundle The NIB bundle.
  */
 - (instancetype)initWithEngine:(FlutterEngine*)engine
-                       nibName:(NSString*)nibNameOrNil
-                        bundle:(NSBundle*)nibBundleOrNil NS_DESIGNATED_INITIALIZER;
+                       nibName:(nullable NSString*)nibName
+                        bundle:(nullable NSBundle*)nibBundle NS_DESIGNATED_INITIALIZER;
 
 /**
  * Initializes a new FlutterViewController and `FlutterEngine` with the specified
  * `FlutterDartProject`.
  *
- * @param projectOrNil The `FlutterDartProject` to initialize the `FlutterEngine` with.
- * @param nibNameOrNil The NIB name to initialize this UIViewController with.
- * @param nibBundleOrNil The NIB bundle.
+ * @param project The `FlutterDartProject` to initialize the `FlutterEngine` with.
+ * @param nibName The NIB name to initialize this UIViewController with.
+ * @param nibBundle The NIB bundle.
  */
-- (instancetype)initWithProject:(FlutterDartProject*)projectOrNil
-                        nibName:(NSString*)nibNameOrNil
-                         bundle:(NSBundle*)nibBundleOrNil NS_DESIGNATED_INITIALIZER;
+- (instancetype)initWithProject:(nullable FlutterDartProject*)project
+                        nibName:(nullable NSString*)nibName
+                         bundle:(nullable NSBundle*)nibBundle NS_DESIGNATED_INITIALIZER;
 
 - (void)handleStatusBarTouches:(UIEvent*)event;
 
@@ -110,7 +122,7 @@ FLUTTER_EXPORT
 
 /**
  * Instructs the Flutter Navigator (if any) to push a route on to the navigation
- * stack.  The setInitialRoute method should be prefered if this is called before the
+ * stack.  The setInitialRoute method should be preferred if this is called before the
  * FlutterViewController has come into view.
  *
  * @param route The name of the route to push to the navigation stack.
@@ -121,6 +133,14 @@ FLUTTER_EXPORT
  * The `FlutterPluginRegistry` used by this FlutterViewController.
  */
 - (id<FlutterPluginRegistry>)pluginRegistry;
+
+/**
+ * True if at least one frame has rendered and the ViewController has appeared.
+ *
+ * This property is reset to false when the ViewController disappears. It is
+ * guaranteed to only alternate between true and false for observers.
+ */
+@property(nonatomic, readonly, getter=isDisplayingFlutterUI) BOOL displayingFlutterUI;
 
 /**
  * Specifies the view to use as a splash screen. Flutter's rendering is asynchronous, so the first
@@ -156,6 +176,16 @@ FLUTTER_EXPORT
  */
 @property(weak, nonatomic, readonly) FlutterEngine* engine;
 
+/**
+ * The `FlutterBinaryMessenger` associated with this FlutterViewController (used for communicating
+ * with channels).
+ *
+ * This is just a convenient way to get the |FlutterEngine|'s binary messenger.
+ */
+@property(nonatomic, readonly) NSObject<FlutterBinaryMessenger>* binaryMessenger;
+
 @end
+
+NS_ASSUME_NONNULL_END
 
 #endif  // FLUTTER_FLUTTERVIEWCONTROLLER_H_
